+++ projects/habitat_ovmm/evaluator.py
@@ -21,7 +21,8 @@ from home_robot.agent.ovmm_agent.ovmm_agent import OpenVocabManipAgent
 from home_robot_sim.env.habitat_ovmm_env.habitat_ovmm_env import (
     HabitatOpenVocabManipEnv,
 )
-
+from metrics_utils import get_stats_from_episode_metrics
+import pandas as pd
 
 def create_ovmm_env_fn(config):
     """Create habitat environment using configsand wrap HabitatOpenVocabManipEnv around it. This function is used by VectorEnv for creating the individual environments"""
@@ -91,6 +92,23 @@ class OVMMEvaluator(PPOTrainer):
             json.dump(aggregated_metrics, f, indent=4)
         with open(f"{self.results_dir}/episode_results.json", "w") as f:
             json.dump(episode_metrics, f, indent=4)
+
+
+    def summarize_metrics(self, episode_metrics):
+        # convert to a dataframe
+        episode_metrics_df = pd.DataFrame.from_dict(episode_metrics, orient="index")
+        episode_metrics_df["start_idx"] = 0
+        stats = get_stats_from_episode_metrics(episode_metrics_df)
+        return stats
+
+    def print_summary(self, summary):
+        print("="*50)
+        print('Averaged metrics')
+        print("="*50)
+        for k, v in summary.items():
+            print(f"{k}: {v}")
+        print("="*50)
+
 
     def _eval(
         self,
@@ -177,3 +195,5 @@ class OVMMEvaluator(PPOTrainer):
 
         envs.close()
         self.write_results(episode_metrics)
+        summary = self.summarize_metrics(episode_metrics)
+        self.print_summary(summary)
